+++ Shared/Kiwix-Bridging-Header.h
@@ -5,3 +5,5 @@
 #import "ZimMultiReader.h"
 #import "OPDSStreamParser.h"
 #import "ZimFileMetaData.h"
+#import "SearchOperation.h"
+#import "SearchResult.h"