+++ iOS/Controller/Search/SearchResultsListController.swift
@@ -87,20 +87,8 @@ class SearchResultsListController: UITableViewController {
         } catch {}
         
         if let snippet = result.snippet {
-            cell.detailLabel.text = snippet
-        } else if let attributedSnippet = result.attributedSnippet {
-            if #available(iOS 13.0, *) {
-                let mutableSnippet = NSMutableAttributedString(attributedString: attributedSnippet)
-                mutableSnippet.addAttributes(
-                    [NSAttributedString.Key.foregroundColor: UIColor.label],
-                    range: NSRange(location: 0, length: mutableSnippet.length)
-                )
-                cell.detailLabel.attributedText = mutableSnippet
-            } else {
-                cell.detailLabel.attributedText = attributedSnippet
-            }
-        } else {
-            cell.detailLabel.text = nil
+            cell.detailLabel.attributedText = snippet
+            cell.detailLabel.numberOfLines = 4
         }
         
         return cell
@@ -115,7 +103,7 @@ class SearchResultsListController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
-        if results[indexPath.row].snippet != nil || results[indexPath.row].attributedSnippet != nil {
+        if results[indexPath.row].snippet != nil {
             return traitCollection.horizontalSizeClass == .regular ? 120 : 190
         } else {
             return 44